% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadShinyMlr.R
\name{downloadShinyMlr}
\alias{downloadShinyMlr}
\title{Download an instance of shinyMlr to your machine}
\usage{
downloadShinyMlr(path = getwd())
}
\arguments{
\item{path}{[\code{character(1)}]\cr
The directory to download the application to.}
}
\description{
Download an instance of shinyMlr to your machine
}
\examples{
\dontrun{
    downloadShinyMlr()
}
}
\seealso{
runShinyMlr
}
