% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyMlr.R
\name{runShinyMlr}
\alias{runShinyMlr}
\title{Run a local instance of shinyMlr}
\usage{
runShinyMlr(path = getwd(), ...)
}
\arguments{
\item{path}{[\code{character(1)}]\cr
The directory containing the application folder.}

\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyMlr
}
\examples{
\dontrun{
  library(shinyMlr)
  downloadShinyMlr()
   runShinyMlr()
}
}
\seealso{
downloadShinyMlr
}
