% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyMlr.R
\name{runShinyMlr}
\alias{runShinyMlr}
\title{Run a local instance of shinyMlr}
\usage{
runShinyMlr(path = getwd(), ...)
}
\arguments{
\item{path}{The directory containing the intRo folder}

\item{...}{Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyMlr
}
\examples{
\dontrun{
  library(shinyMlr)
  downloadShinyMlr()
   runShinyMlr()
}
}
