% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyMlr.R
\name{runShinyMlr}
\alias{runShinyMlr}
\title{Run shinyMlr}
\usage{
runShinyMlr(...)
}
\arguments{
\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyMlr.
}
\examples{
\dontrun{
  runShinyMlr()
}
}
